requirements.txt
amqp==2.3.2
appnope==0.1.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.4
celery==4.2.1
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
decorator==4.3.0
Django==2.1.7
django-bootstrap-breadcrumbs==0.9.1
django-bootstrap-modal-forms==1.4.1
django-celery-beat==1.3.0
django-celery-results==1.0.2
django-cors-headers==2.4.0
django-datatables-view==1.17.0
django-extensions==2.1.5
django-filter==2.0.0
django-fontawesome==1.0
django-geojson==2.11.0
django-leaflet==0.24.0
django-mathfilters==0.4.0
django-templated-mail==1.1.1
django-timezone-field==3.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-datatables==0.4.1
djoser==1.2.1
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
Fiona==1.8.6
flower==0.9.3
geographiclib==1.49
geomet==0.2.0.post2
geopy==1.17.0
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
jobtastic==2.0.0
jsonfield==2.0.2
jsonschema==3.0.1
kombu==4.2.1
Markdown==3.0.1
munch==2.3.2
numpy==1.15.2
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.0.0
pprint==0.1
promise==2.2.1
prompt-toolkit==2.0.7
psutil==3.4.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.3.1
pyrsistent==0.14.11
python-crontab==2.3.5
python-dateutil==2.7.3
pytz==2018.5
PyYAML==4.2b1
redis==2.10.6
rfc3987==1.3.8
Shapely==1.6.4.post2
simplejson==3.16.0
six==1.11.0
topojson==1.0.1+2.g581c6b4
tornado==5.1.1
traitlets==4.3.2
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
